shop_and_pet/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.32
